Rename `grakn-engine-client` to `grakn-client`
