+++ grakn-client/pom.xml
@@ -27,7 +27,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>grakn-engine-client</artifactId>
+    <artifactId>grakn-client</artifactId>
 
     <dependencies>
         <dependency>
