+++ grakn-client/src/main/java/ai/grakn/client/LoaderClient.java
@@ -16,7 +16,7 @@
  * along with Grakn. If not, see <http://www.gnu.org/licenses/gpl.txt>.
  */
 
-package ai.grakn.engine.client;
+package ai.grakn.client;
 
 import ai.grakn.engine.TaskStatus;
 import ai.grakn.graql.InsertQuery;
