+++ grakn-graql-shell/pom.xml
@@ -57,7 +57,7 @@
         </dependency>
         <dependency>
             <groupId>ai.grakn</groupId>
-            <artifactId>grakn-engine-client</artifactId>
+            <artifactId>grakn-client</artifactId>
             <version>${project.version}</version>
         </dependency>
     </dependencies>
