+++ grakn-migration/base/pom.xml
@@ -36,7 +36,7 @@
         </dependency>
         <dependency>
             <groupId>ai.grakn</groupId>
-            <artifactId>grakn-engine-client</artifactId>
+            <artifactId>grakn-client</artifactId>
             <version>${project.version}</version>
         </dependency>
     </dependencies>
