+++ grakn-migration/base/src/main/java/ai/grakn/migration/base/AbstractMigrator.java
@@ -19,7 +19,7 @@
 package ai.grakn.migration.base;
 
 import ai.grakn.engine.TaskStatus;
-import ai.grakn.engine.client.LoaderClient;
+import ai.grakn.client.LoaderClient;
 import ai.grakn.exception.GraqlTemplateParsingException;
 import ai.grakn.graql.Graql;
 import ai.grakn.graql.InsertQuery;
